Activity refactor

+++ src/main/java/com/asialocalguide/gateway/viator/dto/ViatorActivitySearchDTO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public record ViatorActivitySearchDTO(Filtering filtering, Sorting sorting, Pagination pagination, String currency) {
 
-  public record Filtering(Long destination, List<Long> tags, LocalDate startDate, LocalDate endDate, Range rating) {}
+  public record Filtering(Long destination, List<Integer> tags, LocalDate startDate, LocalDate endDate, Range rating) {}
 
   public record Range(Integer from, Integer to) {}
